Refactor(redirect): 구조 재 정의
