+++ src/app/routes/path.ts
@@ -1,7 +1,7 @@
 export const RouterPath = {
   ROOT: '/',
   LOGIN: '/login',
-  REDIRECT: '/redirection',
+  SIGNUP: '/signup',
   REGISTER: '/register',
   GUARD_MYPAGE: `/guard/mypage`,
   GUARD_GUIDELINE: `/guard/guideline`,
