+++ src/pages/common/register/RegisterPage.tsx
@@ -2,11 +2,9 @@ import { useState } from 'react';
 import { useForm } from 'react-hook-form';
 
 import useRegister from './api/hooks/useRegister';
-import { RegisterFields } from './components/register-fields';
-import { RegisterType } from './components/register-type';
-import { Tos } from './components/tos';
+import { RegisterFields, RegisterType, Tos } from './components';
 import { FormValues } from './types';
-import { BasicButton } from '@/shared/components/common/button';
+import { BasicButton } from '@/shared/components';
 import { Divider } from '@chakra-ui/react';
 import styled from '@emotion/styled';
