+++ src/pages/common/register/components/index.ts
@@ -1,0 +1,6 @@
+export * from './register-fields';
+export * from './register-fields/form-fields';
+export * from './register-type';
+export * from './register-type/type-button';
+export * from './tos';
+export * from './tos/content';
