+++ src/pages/common/register/components/register-fields/form-fields/index.ts
@@ -1,0 +1,1 @@
+export { FormField } from './FormField';
