+++ src/pages/common/register/components/register-fields/index.ts
@@ -1,0 +1,1 @@
+export { RegisterFields } from './RegisterFields';
