+++ src/pages/common/register/components/register-type/index.ts
@@ -1,0 +1,1 @@
+export { RegisterType } from './RegisterType';
