+++ src/pages/common/register/components/register-type/type-button/TypeButton.tsx
