+++ src/pages/common/register/components/register-type/type-button/index.ts
@@ -1,0 +1,1 @@
+export { TypeButton } from './TypeButton';
