+++ src/pages/common/register/components/tos/Tos.tsx
