+++ src/pages/common/register/components/tos/content/TosContent.tsx
