+++ src/pages/common/register/components/tos/content/index.ts
@@ -1,0 +1,1 @@
+export { TosContent } from './TosContent';
