+++ src/pages/common/register/components/tos/index.ts
@@ -1,0 +1,1 @@
+export { Tos } from './Tos';
