+++ src/shared/components/common/button/index.ts
@@ -1,0 +1,1 @@
+export { BasicButton } from './BasicButton';
