+++ src/shared/components/common/index.ts
@@ -1,0 +1,2 @@
+export * from './button';
+export * from './layout';
