+++ src/shared/components/features/index.ts
@@ -1,0 +1,3 @@
+export * from './header';
+export * from './mypage';
+export * from './notice';
