+++ src/shared/components/features/mypage/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useToggleDetailBox } from './useToggleDetailBox';
