+++ src/shared/components/features/mypage/index.ts
@@ -1,0 +1,3 @@
+export * from './hooks';
+export * from './point-box';
+export * from './use-detail';
