+++ src/shared/components/features/mypage/point-box/PointBox.tsx
