+++ src/shared/components/features/mypage/point-box/index.ts
@@ -1,0 +1,1 @@
+export { default as PointBox } from './PointBox';
