+++ src/shared/components/features/mypage/use-detail/index.ts
@@ -1,0 +1,1 @@
+export { default as UseDetailBox } from './UseDetailBox';
