+++ src/shared/components/features/notice/Notice.tsx
