+++ src/shared/components/features/notice/index.ts
@@ -1,0 +1,1 @@
+export { default as Notice } from './Notice';
